64bit XADMaster also some small fixes to make SC compile without warnings. SC is now a 64bit app.

+++ Custom Windows/Bezel/TSSTBezelWindow.m
@@ -15,7 +15,7 @@
 @implementation TSSTBezelWindow
 
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
 {
     self = [super initWithContentRect: contentRect styleMask: NSBorderlessWindowMask | NSNonactivatingPanelMask backing: bufferingType defer: flag];
     if(self)
